feat: Refactor module imports in gwkokab models
