+++ gwkokab/_src/models/__init__.py
@@ -12,7 +12,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-
 from .models import (
     BrokenPowerLawMassModel as BrokenPowerLawMassModel,
     GaussianSpinModel as GaussianSpinModel,
