+++ gwkokab/models/utils.py
@@ -13,10 +13,8 @@
 #  limitations under the License.
 
 
-from gwkokab._src.models.utils.jointdistribution import (
-    JointDistribution as JointDistribution,
-)
-from gwkokab._src.models.utils.wrappers import (
+from gwkokab._src.models.utils import JointDistribution as JointDistribution
+from gwkokab._src.models.wrappers import (
     add_log_factor as add_log_factor,
     gwk_magazine as gwk_magazine,
 )
